+++ src/sardana/macroserver/macros/env.py
@@ -87,6 +87,8 @@ class setvo(Macro):
     - **ShowCtrlAxis**: used by macro wm, pwm and wa. Default value ``False``
     - **PosFormat**: used by macro wm, pwm and wa. Default value ``-1``
     - **OutputBlock**: used by scan macros. Default value ``False``
+    - **DescriptionLength**: used by lsdef. Default value ``60``
+
 
     """
 
@@ -112,6 +114,7 @@ class usetvo(Macro):
     - **ShowCtrlAxis**: used by macro wm, pwm and wa. Default value ``False``
     - **PosFormat**: used by macro wm, pwm and wa. Default value ``-1``
     - **OutputBlock**: used by scan macros. Default value ``False``
+    - **DescriptionLength**: used by lsdef. Default value ``60``
 
     """