+++ src/sardana/macroserver/macros/hkl.py
@@ -496,7 +496,7 @@ class pa(Macro, _diffrac):
                 self.output("  %d%s Reflection (index %d): " %
                             (nb_ref + 1, sf, ref[0]))
                 #self.output("    Affinement, Relevance : %d %d" % (ref[4], ref[5]))
-                if len(ref) > 10:
+                if len(ref) > 10 and len(self.angle_names) == 6:
                     self.output("    %s %s %s %s %s %s: %s %s %s %s %s %s" %
                                 (self.angle_names[5], self.angle_names[1],
                                  self.angle_names[2], self.angle_names[3],
@@ -507,6 +507,20 @@ class pa(Macro, _diffrac):
                                  _diffrac.fl(self, str(ref[9])),
                                  _diffrac.fl(self, str(ref[10])),
                                  _diffrac.fl(self, str(ref[6]))))
+                elif len(ref) > 10 and len(self.angle_names) == 7:
+                    self.output(
+                        "    %s %s %s %s %s %s %s: %s %s %s %s %s %s %s" %
+                        (self.angle_names[0], self.angle_names[1],
+                         self.angle_names[2], self.angle_names[3],
+                         self.angle_names[4], self.angle_names[5],
+                         self.angle_names[6],
+                         _diffrac.fl(self, str(ref[6])),
+                         _diffrac.fl(self, str(ref[7])),
+                         _diffrac.fl(self, str(ref[8])),
+                         _diffrac.fl(self, str(ref[9])),
+                         _diffrac.fl(self, str(ref[10])),
+                         _diffrac.fl(self, str(ref[11])),
+                         _diffrac.fl(self, str(ref[12]))))
                 else:
                     self.output("    %s %s %s %s: %s %s %s %s" %
                                 (self.angle_names[0], self.angle_names[1],
@@ -657,19 +671,19 @@ class wh(Macro, _diffrac):
                         (str_pos1, str_pos2, str_pos3, str_pos4))
         elif self.nb_motors == 7:
             str_pos1 = "%7.5f" % self.getDevice(
-                self.angle_device_names[self.angles_names[0]]).Position
+                self.angle_device_names[self.angle_names[0]]).Position
             str_pos2 = "%7.5f" % self.getDevice(
-                self.angle_device_names[self.angles_names[1]]).Position
+                self.angle_device_names[self.angle_names[1]]).Position
             str_pos3 = "%7.5f" % self.getDevice(
-                self.angle_device_names[self.angles_names[2]]).Position
+                self.angle_device_names[self.angle_names[2]]).Position
             str_pos4 = "%7.5f" % self.getDevice(
-                self.angle_device_names[self.angles_names[3]]).Position
+                self.angle_device_names[self.angle_names[3]]).Position
             str_pos5 = "%7.5f" % self.getDevice(
-                self.angle_device_names[self.angles_names[4]]).Position
+                self.angle_device_names[self.angle_names[4]]).Position
             str_pos6 = "%7.5f" % self.getDevice(
-                self.angle_device_names[self.angles_names[5]]).Position
+                self.angle_device_names[self.angle_names[5]]).Position
             str_pos7 = "%7.5f" % self.getDevice(
-                self.angle_device_names[self.angles_names[6]]).Position
+                self.angle_device_names[self.angle_names[6]]).Position
             self.output("%10s %11s %12s %11s %11s %11s %11s" %
                         (self.angle_names[0],
                          self.angle_names[1],