+++ src/sardana/tango/pool/TwoDExpChannel.py
@@ -33,7 +33,7 @@ import sys
 import time
 
 from PyTango import DevFailed, DevVoid, DevString, DevState, AttrQuality, \
-    Except, READ, SCALAR
+    Except, READ, SCALAR, READ_WRITE
 
 from taurus.core.util.log import DebugIt
 
@@ -42,17 +42,17 @@ from sardana.sardanaattribute import SardanaAttribute
 from sardana.pool.controller import TwoDController, MaxDimSize, Type
 from sardana.tango.core.util import to_tango_type_format, exception_str
 
-from sardana.tango.pool.PoolDevice import PoolElementDevice, \
-    PoolElementDeviceClass
+from sardana.tango.pool.PoolDevice import PoolTimerableDevice, \
+    PoolTimerableDeviceClass
 
 
-class TwoDExpChannel(PoolElementDevice):
+class TwoDExpChannel(PoolTimerableDevice):
 
     def __init__(self, dclass, name):
-        PoolElementDevice.__init__(self, dclass, name)
+        PoolTimerableDevice.__init__(self, dclass, name)
 
     def init(self, name):
-        PoolElementDevice.init(self, name)
+        PoolTimerableDevice.init(self, name)
 
     def get_twod(self):
         return self.element
@@ -64,14 +64,14 @@ class TwoDExpChannel(PoolElementDevice):
 
     @DebugIt()
     def delete_device(self):
-        PoolElementDevice.delete_device(self)
+        PoolTimerableDevice.delete_device(self)
         twod = self.twod
         if twod is not None:
             twod.remove_listener(self.on_twod_changed)
 
     @DebugIt()
     def init_device(self):
-        PoolElementDevice.init_device(self)
+        PoolTimerableDevice.init_device(self)
         twod = self.twod
         if twod is None:
             full_name = self.get_full_name()
@@ -106,6 +106,7 @@ class TwoDExpChannel(PoolElementDevice):
 
         timestamp = time.time()
         name = event_type.name.lower()
+        name = name.replace('_', '')  # for integration_time events
 
         try:
             attr = self.get_attribute_by_name(name)
@@ -127,8 +128,18 @@ class TwoDExpChannel(PoolElementDevice):
                 else:
                     value = event_value.value
                 timestamp = event_value.timestamp
+            else:
+                value = event_value
 
-            if name == "value":
+            if name == "timer" and value is None:
+                value = "None"
+            elif name == "datasource" and value is None:
+                full_name = self.get_full_name()
+                # for Taurus 3/4 compatibility
+                if not full_name.startswith("tango://"):
+                    full_name = "tango://{0}".format(full_name)
+                value = "{0}/value".format(full_name)
+            elif name == "value":
                 state = self.twod.get_state()
                 if state == State.Moving:
                     quality = AttrQuality.ATTR_CHANGING
@@ -147,7 +158,7 @@ class TwoDExpChannel(PoolElementDevice):
         cache_built = hasattr(self, "_dynamic_attributes_cache")
 
         std_attrs, dyn_attrs = \
-            PoolElementDevice.get_dynamic_attributes(self)
+            PoolTimerableDevice.get_dynamic_attributes(self)
 
         if not cache_built:
             # For value attribute, listen to what the controller says for data
@@ -204,7 +215,7 @@ _DFT_VALUE_TYPE, _DFT_VALUE_FORMAT = to_tango_type_format(
     _DFT_VALUE_INFO[Type], DataFormat.TwoD)
 
 
-class TwoDExpChannelClass(PoolElementDeviceClass):
+class TwoDExpChannelClass(PoolTimerableDeviceClass):
 
     #    Class Properties
     class_property_list = {
@@ -213,29 +224,29 @@ class TwoDExpChannelClass(PoolElementDeviceClass):
     #    Device Properties
     device_property_list = {
     }
-    device_property_list.update(PoolElementDeviceClass.device_property_list)
+    device_property_list.update(PoolTimerableDeviceClass.device_property_list)
 
     #    Command definitions
     cmd_list = {
         'Start':   [[DevVoid, ""], [DevVoid, ""]],
     }
-    cmd_list.update(PoolElementDeviceClass.cmd_list)
+    cmd_list.update(PoolTimerableDeviceClass.cmd_list)
 
     #    Attribute definitions
     attr_list = {
         'DataSource': [[DevString, SCALAR, READ]],
     }
-    attr_list.update(PoolElementDeviceClass.attr_list)
+    attr_list.update(PoolTimerableDeviceClass.attr_list)
 
     standard_attr_list = {
         'Value': [[_DFT_VALUE_TYPE, _DFT_VALUE_FORMAT, READ,
                    _DFT_VALUE_MAX_SHAPE[0], _DFT_VALUE_MAX_SHAPE[1]],
                   {'abs_change': '1.0', }],
     }
-    standard_attr_list.update(PoolElementDeviceClass.standard_attr_list)
+    standard_attr_list.update(PoolTimerableDeviceClass.standard_attr_list)
 
     def _get_class_properties(self):
-        ret = PoolElementDeviceClass._get_class_properties(self)
+        ret = PoolTimerableDeviceClass._get_class_properties(self)
         ret['Description'] = "2D device class"
-        ret['InheritedFrom'].insert(0, 'PoolElementDevice')
+        ret['InheritedFrom'].insert(0, 'PoolTimerableDevice')
         return ret